Fix: move custom components to main directory
