+++ web/src/components/SignUpCompany/index.tsx
