+++ web/src/components/index.ts
@@ -56,3 +56,9 @@ export * from './Info';
 export * from './LockedComponent';
 export * from './WarningMessage';
 export * from './SearchBox';
+export * from './SignUpCompany';
+export * from './SignUpSelection';
+export * from './HeaderProfile';
+export * from './CompanyAccount';
+export * from './DownloadBalance';
+export * from './CalendarPicker';
