+++ web/src/components/style.pcss
@@ -49,3 +49,9 @@
 @import "./Beneficiaries/BeneficiariesCrypto/SelectBeneficiariesCrypto.pcss";
 @import "./WarningMessage/WarningMessage.pcss";
 @import "./SearchBox/SearchBox.pcss";
+@import './SignUpCompany/SignUpCompany.pcss';
+@import './SignUpSelection/SignUpSelection.pcss';
+@import './HeaderProfile/HeaderProfile.pcss';
+@import './CompanyAccount/CompanyAccount.pcss';
+@import './DownloadBalance/DownloadBalance.pcss';
+@import './CalendarPicker/CalendarPicker.pcss';
