--- web/src/custom/components/index.tsx
@@ -1,6 +1,0 @@
-export * from './SignUpCompany';
-export * from './SignUpSelection';
-export * from './HeaderProfile';
-export * from './CompanyAccount';
-export * from './DownloadBalance';
-export * from './CalendarPicker';
