--- web/src/custom/components/style.pcss
@@ -1,6 +1,0 @@
-@import './SignUpCompany/SignUpCompany.pcss';
-@import './SignUpSelection/SignUpSelection.pcss';
-@import './HeaderProfile/HeaderProfile.pcss';
-@import './CompanyAccount/CompanyAccount.pcss';
-@import './DownloadBalance/DownloadBalance.pcss';
-@import './CalendarPicker/CalendarPicker.pcss';
