--- web/src/custom/containers/Wallets/EstimatedValue/EstimatedValue.pcss
@@ -1,46 +1,0 @@
-.pg-estimated-value {
-  background: var(--body-background-color);
-  border-radius: 3px;
-  border-bottom-left-radius: 0;
-  border-bottom-right-radius: 0;
-  display: flex;
-  flex-direction: column;
-  margin: 24px 0 0;
-  max-width: 1368px;
-  min-height: 101px;
-  padding: 24px;
-  width: 100%;
-
-  &__title {
-    color: var(--secondary-contrast-cta-color);
-    font-size: 12px;
-    margin-bottom: 2px;
-  }
-
-  &__container {
-    display: flex;
-    flex-direction: row;
-
-    .value-container {
-      align-items: center;
-      color: var(--primary-text-color);
-      display: flex;
-      margin-right: 48px;
-
-      .value {
-        color: var(--primary-text-color);
-        font-size: 22px;
-      }
-
-      .value-sign {
-        background: var(--primary-cta-color-level-1);
-        border-radius: 3px;
-        color: var(--primary-cta-color);
-        font-size: 14px;
-        letter-spacing: 0.0357143em;
-        margin-left: 6px;
-        padding: 3px;
-      }
-    }
-  }
-}
