--- web/src/custom/containers/Wallets/OrganizationOverview/OrganizationOverview.pcss
@@ -1,116 +1,0 @@
-
-.cr-wallets-organization-table {
-  &__title {
-    color: var(--primary-text-color);
-    margin-bottom: 18px;
-  }
-
-  &__table {
-    .cr-table {
-      &__head th {
-        border-bottom: 1px solid var(--divider-color);
-        color: var(--secondary-contrast-cta-color);
-        font-size: 14px;
-        font-weight: normal;
-        padding: 12px 24px;
-        text-align: right !important;
-
-        &:last-child {
-          padding-left: 48px;
-          width: unset;
-        }
-      }
-
-      &__head-row {
-        border-bottom: 1px solid var(--table-border-color);
-      }
-
-      &__body tr {
-        border-bottom: 1px solid var(--table-border-color);
-        height: 48px;
-      }
-
-      &__body tr td {
-        border-bottom: 1px solid var(--divider-color);
-        color: var(--primary-text-color);
-        font-size: 14px !important;
-        letter-spacing: 0.1px;
-        padding: 16px 24px !important;
-        text-align: right;
-      }
-
-      &__head th,
-      &__body tr td {
-        text-align: left !important;
-
-        &:first-child,
-        &:nth-child(2) {
-          text-align: left !important;
-        }
-
-        &:last-child {
-          padding-left: 48px !important;
-          width: 20%;
-        }
-      }
-
-      &--no-data {
-        text-align: center !important;
-      }
-    }
-  }
-
-  &__wallet {
-    display: flex;
-    flex-direction: row;
-
-    &--icon {
-      margin-right: 12px;
-
-      img {
-        height: 30px;
-      }
-    }
-
-    &--description {
-      align-items: center;
-      display: flex;
-
-      .bold {
-        font-size: 16px;
-        font-weight: 600;
-        letter-spacing: 0.15px;
-        margin-right: 6px;
-      }
-
-      .secondary {
-        color: var(--secondary-contrast-cta-color);
-        font-size: 14px;
-        letter-spacing: 0.1px;
-      }
-    }
-  }
-
-  &__button-wrapper {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-  }
-
-  .btn-secondary {
-    background: none;
-    border: 1px solid var(--primary-text-color-level-1);
-    border-radius: unset;
-    color: var(--primary-text-color-level-7);
-    font-size: 12px;
-    height: 36px;
-    letter-spacing: 0.04em;
-    margin-left: 12px;
-    padding: 0 12px;
-    width: 103px;
-
-    &:first-child {
-      margin-left: 0;
-    }
-  }
-}
