--- web/src/custom/containers/style.pcss
@@ -1,2 +1,0 @@
-@import './Wallets/OrganizationOverview/OrganizationOverview.pcss';
-@import './Wallets/EstimatedValue/EstimatedValue.pcss';
