+++ web/src/custom/styles/style.pcss
@@ -1,4 +1,3 @@
 @import "./fonts.pcss";
 @import "./themes/style.pcss";
 @import "./colors.pcss";
-@import "../components/style.pcss";
